mask support for templates
Provides optional support for a masking the template during search, and
a sample mask for the colour_display_template. The area taken by a max
size square brush is masked off as this should be the largest area of
that template that can change depending on current brush size/selection.
Also useful for brush and tools menu, though no mask is provided.